Player controller (XR Rig) added
Installed the XR Interaction Toolkit within the Package Manager. Created some additional top-level folders to organize project files. Created XR Rig prefab (device based), continuous movement & turning has been configured to work with the right controller's joystick. Movement configured to move based on the player's in-game orientation and not based on the headset's rotation. Needs to be tested with the Quest 2, controls should be the same.